// src/app/offset/models/offset-technical-details.ts
export class OffsetTechnicalDetails {
    offsettechnicalid: any;
    screenprintid: any;
    sampleid: any;
    laminationid: any;
    thermal: any;
    coldglue: any;
    eyeletid: any;
    stringid: any;
    tagpinid: any;
    cuttingid: any;
    embossid: any;
    debossid: any;
    foilingid: any;
    boardpasting: any;
    windowid: any;
    windowwithpetid: any;
    hangerid: any;
    hookid: any;
    holeid: any;
    perforationid: any;
    bothsidetapeid: any;
    magnetid: any;
    stickerattachingid: any;
    rfidattachingid: any;
    folderglueingid: any;
    creezing: any;
    uv: any;
    noofups: any;
    description1: any;
    description2: any;
    description3: any;
    description4: any;
    description5: any;
    description6: any;
    description7: any;
    description8: any;
    description9: any;
    description10: any;
    description11: any;
    description12: any;
    description13: any;
    description14: any;
    description15: any;
    description16: any;
    description17: any;
    description18: any;
    description19: any;
    description20: any;
    description21: any;
    constructor() {
        this.sampleid = 0;
        this.screenprintid = 0;
        this.laminationid = 0;
        this.thermal = 0;
        this.coldglue = 0;
        this.eyeletid = 0;
        this.stringid = 0;
        this.tagpinid = 0;
        this.cuttingid = 0;
        this.embossid = 0;
        this.debossid = 0;
        this.foilingid = 0;
        this.boardpasting = 0;
        this.windowid = 0;
        this.windowwithpetid = 0;
        this.hangerid = 0;
        this.hookid = 0;
        this.holeid = 0;
        this.perforationid = 0;
        this.bothsidetapeid = 0;
        this.magnetid = 0;
        this.stickerattachingid = 0;
        this.rfidattachingid = 0;
        this.folderglueingid = 0;
        this.creezing = 0;
        this.uv = 0;
       this.description1 = '';
       this.description2 = '';
       this.description3 = '';
       this.description4 = '';
       this.description5 = '';
       this.description6 = '';
       this.description7 = '';
       this.description8 = '';
       this.description9 = '';
       this.description10 = '';
       this.description11 = '';
       this.description12 = '';
       this.description13 = '';
       this.description14 = '';
       this.description15 = '';
       this.description16 = '';
       this.description17 = '';
       this.description18 = '';
       this.description19 = '';
       this.description20 = '';
       this.description21 = '';
    }
}